cmd/k8s-operator: add reconciler for Tailnet resource
This commit adds the new reconciler for the Tailnet resource in a new
location that we want to put all future reconcilers in and eventually
refactor existing reconcilers into.

The Tailnet resource represents OAuth credentials for a single tailnet and
uses status fields to indicate the validity of the credentials. Validity
is determined via expected fields existing (client_id and client_secret)
and API calls for devices, keys & services being possible using them.

Subsequent resources will be updated to refer to a Tailnet as the source
of their oauth credentials, or fall back to the configured default.

Fixes: https://github.com/tailscale/corp/issues/34767
Signed-off-by: David Bond <davidsbond93@gmail.com>